Split out the form into a Vue Component
I decided to do this as I felt that I was likely going to add some
pretty messy JavaScript into the page, JavaScript that wasn't the
easiest to maintain or develop. Secondly, I wanted to keep on learning
about Vue.js, my preferred front end framework. I feel that by using it
that the code will be much more flexible, powerful, and maintainable.